v6/unidirectional_sse/backend/requirements.txt
fastapi==0.115.12
httpx==0.28.1
uvicorn==0.34.3
PyJWT==2.8.0
